requirements.txt
# Core dependencies for IT Management Automation
python-dateutil>=2.8.0
pyyaml>=6.0
tabulate>=0.9.0
